requirements.txt
discord
PyPDF2
python-dotenv